Database Design Logical Phase
This was the second step in creating my database. This step shows logical knowledge of creating a database. We were tasked with creating a Visio diagram to depict how our information would be organized using tables, primary keys, and foreign keys. This is the layout on which we based our Healthcare Form Design.sql file on.